+++ Sources/GameMath/Color.swift
@@ -356,9 +356,7 @@ public extension Color {
     static let yellow: Color        = #colorLiteral(red: 1.0, green: 1.0, blue: 0.0, alpha: 1.0)
     static let orange: Color        = #colorLiteral(red: 1.0, green: 0.64453125, blue: 0.0, alpha: 1.0)
     static let purple: Color        = #colorLiteral(red: 0.5, green: 0.0, blue: 0.5, alpha: 1.0)
-    
-    internal static let stregasgateBackground: Color = #colorLiteral(red: 0.094117634, green: 0.0941176638, blue: 0.094117634, alpha: 1)
-    
+        
     static let vertexColors = Color(red: -1001, green: -2002, blue: -3003, alpha: -4004)
     static let defaultDiffuseMapColor = Color(red: 0.5, green: 0.5, blue: 0.5, alpha: 1)
     static let defaultNormalMapColor = Color(red: 0.5, green: 0.5, blue: 1.0, alpha: 1.0)
